Reduce size of RoomDetailFragment
