+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/RoomDetailFragment.kt
@@ -181,6 +181,7 @@ class RoomDetailFragment @Inject constructor(
 
     private lateinit var sharedActionViewModel: MessageSharedActionViewModel
     private lateinit var layoutManager: LinearLayoutManager
+    private lateinit var jumpToBottomViewVisibilityManager: JumpToBottomViewVisibilityManager
     private var modelBuildListener: OnModelBuildFinishedListener? = null
 
     private lateinit var attachmentsHelper: AttachmentsHelper
@@ -312,6 +313,13 @@ class RoomDetailFragment @Inject constructor(
                 }
             }
         }
+
+        jumpToBottomViewVisibilityManager = JumpToBottomViewVisibilityManager(
+                jumpToBottomView,
+                debouncer,
+                recyclerView,
+                layoutManager
+        )
     }
 
     private fun setupJumpToReadMarkerView() {
@@ -474,34 +482,10 @@ class RoomDetailFragment @Inject constructor(
             it.dispatchTo(scrollOnNewMessageCallback)
             it.dispatchTo(scrollOnHighlightedEventCallback)
             updateJumpToReadMarkerViewVisibility()
-            maybeShowJumpToBottomViewVisibilityWithDelay()
+            jumpToBottomViewVisibilityManager.maybeShowJumpToBottomViewVisibilityWithDelay()
         }
         timelineEventController.addModelBuildListener(modelBuildListener)
         recyclerView.adapter = timelineEventController.adapter
-
-        recyclerView.addOnScrollListener(object : RecyclerView.OnScrollListener() {
-            override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
-                debouncer.cancel("jump_to_bottom_visibility")
-
-                val scrollingToPast = dy < 0
-
-                if (scrollingToPast) {
-                    jumpToBottomView.hide()
-                } else {
-                    maybeShowJumpToBottomViewVisibility()
-                }
-            }
-
-            override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {
-                when (newState) {
-                    RecyclerView.SCROLL_STATE_IDLE     -> {
-                        maybeShowJumpToBottomViewVisibilityWithDelay()
-                    }
-                    RecyclerView.SCROLL_STATE_DRAGGING,
-                    RecyclerView.SCROLL_STATE_SETTLING -> Unit
-                }
-            }
-        })
 
         timelineEventController.callback = this
 
@@ -554,21 +538,6 @@ class RoomDetailFragment @Inject constructor(
                 }
                 jumpToReadMarkerView.isVisible = showJumpToUnreadBanner
             }
-        }
-    }
-
-    private fun maybeShowJumpToBottomViewVisibilityWithDelay() {
-        debouncer.debounce("jump_to_bottom_visibility", 250, Runnable {
-            maybeShowJumpToBottomViewVisibility()
-        })
-    }
-
-    private fun maybeShowJumpToBottomViewVisibility() {
-        Timber.v("First visible: ${layoutManager.findFirstCompletelyVisibleItemPosition()}")
-        if (layoutManager.findFirstVisibleItemPosition() != 0) {
-            jumpToBottomView.show()
-        } else {
-            jumpToBottomView.hide()
         }
     }
